separate Status fiels ro it's own component
